fix: change toss_merchant_id column type from INT to BIGINT
토스 가맹점 ID 값이 INT 범위(최대 21억)를 초과하여 DataError 발생.
BigInteger로 변경하고 DB 마이그레이션 추가.

+++ app/models/__init__.py
@@ -75,7 +75,7 @@ class Store(db.Model):
     receipt_header = db.Column(db.Text, nullable=True) # 영수증 머릿말
     receipt_footer = db.Column(db.Text, nullable=True) # 영수증 꼬릿말
     terminal_serial = db.Column(db.String(50), nullable=True)   # 토스 프론트 단말기 시리얼 번호
-    toss_merchant_id = db.Column(db.Integer, nullable=True)      # 토스 가맹점 ID
+    toss_merchant_id = db.Column(db.BigInteger, nullable=True)      # 토스 가맹점 ID
     toss_business_number = db.Column(db.String(20), nullable=True) # 사업자등록번호
     created_at = db.Column(db.DateTime, default=datetime.now)
     last_logged_at = db.Column(db.DateTime, nullable=True)
